requirements.txt
biopython
Pillow
streamlit